Allow table definitions with custom metadata object

+++ tests/db_test.py
@@ -3,9 +3,17 @@ from nose.plugins.attrib import attr
 
 import os
 import shutil
-from shiftcontent.db import Db
+from sqlalchemy import Table
+from sqlalchemy import Column
+from sqlalchemy import Integer
+from sqlalchemy import String
+from sqlalchemy import ForeignKey
+
+
+from shiftcontent import Db
 from shiftcontent import exceptions as x
 from sqlalchemy.engine.base import Engine
+from sqlalchemy.sql.schema import MetaData
 
 @attr('db')
 class DbTest(unittest.TestCase):
@@ -20,17 +28,17 @@ class DbTest(unittest.TestCase):
             os.makedirs(tmp, exist_ok=True)
         return tmp
 
+    @property
     def db_url(self):
         """ Get test db url """
-        url = 'sqlite://{}/test.db'.format(self.tmp)
+        url = 'sqlite:///{}/test.db'.format(self.tmp)
         return url
 
     def tearDown(self):
         """ Cleanup """
         super().tearDown()
         tests = os.path.join(os.getcwd(), 'var', 'data', 'tests')
-        if os.path.exists(tests):
-            shutil.rmtree(tests)
+        if os.path.exists(tests): shutil.rmtree(tests)
 
     # --------------------------------------------------------------------------
 
@@ -55,3 +63,42 @@ class DbTest(unittest.TestCase):
         mock_engine = 'I am an engine'
         db = Db(engine=mock_engine)
         self.assertEquals(mock_engine, db.engine)
+
+    def test_use_custom_meta(self):
+        """ Use custom metadata object """
+        mock_meta = 'I am metadata!'
+        db = Db(self.db_url, meta=mock_meta)
+        self.assertEquals(mock_meta, db.meta)
+
+    def test_create_fresh_metadata(self):
+        """ Creating metadata object on first access """
+        db = Db(self.db_url)
+        meta = db.meta
+        self.assertIsInstance(meta, MetaData)
+
+    def test_creating_tables(self):
+        """ Creating tables"""
+        db_url = self.db_url
+        db = Db(db_url)
+        employees = Table('employees', db.meta,
+            Column('id', Integer, primary_key=True),
+            Column('name', String(128)),
+            Column('full_name', String(128)),
+        )
+
+        addresses = Table('addresses', db.meta,
+            Column('id', Integer, primary_key=True),
+            Column('user_id', None, ForeignKey('employees.id')),
+            Column('email_address', String(256), nullable=False)
+        )
+
+        db.meta.create_all()
+        self.assertIn(employees, db.meta.sorted_tables)
+        self.assertIn(addresses, db.meta.sorted_tables)
+
+
+
+
+
+
+
